requirements.txt
llama-cpp-python
streamlit
vaderSentiment